// app/interfaces/FlightData.ts
interface ChangeIndicator {
    Name: string;
    Value: string;
}

interface MarketingFlightSegment {
    MarketingFlightNumber: string;
    FlightSegDestLocalDateTime: string;
    SourceSystemName: string;
    SourceMessageType: string;
    MarketingCarrierCode: string;
    FlightSegDestAirport: string;
    FlightSegOriginLocalDateTime: string;
    FlightSegOriginLocalDate: string;
    SourceUpdateDateTime: string;
    FlightSegOriginAirport: string;
    SourceMessageDateTime: string;
}

interface CurrentFlightLeg {
    DepartureVariance: string;
    PrevTailNumber: string;
    CurrentArrivalTerminal: string;
    OriginatorFlightIndicator: string;
    ActualDepartureDelayMinutes: string;
    DispatchDesk: string;
    BlueTurnIndicator: string;
    EstimatedDepartureDelayMinutes: string;
    PrevAcftRegNumber: string;
    ActualEquipmentType: string;
    TurnTime: string;
    DoorCloseTime: string;
    DepGate: string;
    EstLocalArrDateTime: string;
    TailNumber: string;
    ActualFlightType: string;
    Status: string;
    ArrProgress: string;
    CurrentArrivalDateTime: string;
    DepProgress: string;
    BaggageCarousel: string;
    EstimatedGroundMinutes: string;
    EstLocalBoardingDateTime: string;
    ActualArrivalDelayMinutes: string;
    CurrentDepStnCode: string;
    RON: string;
    AcftRegNumber: string;
    ActiveAFLSeqNumber: string;
    CurrentDepartureDateTime: string;
    ATCStatus: string;
    FleetLaunchFlightIndicator: string;
    CurrentArrStnCode: string;
    EstimatedArrivalDelayMinutes: string;
    EstLocalDepDateTime: string;
    SourceUpdateDateTime: string | null;
    ActLocalOutDateTime: string;
    TechStopStationCode: string | null;
    DivertedStationCode: string | null;
    PrevDepGate: string | null;
    CurrentDepartureTerminal: string;
    PrevArrGate: string | null;
    ArrGate: string | null;
    ActLocalOffDateTime: string | null;
    ActLocalOnDateTime: string | null;
    ActLocalInDateTime: string | null;
    ActualBlockMinutes: string | null;
    ActualFlightTimeMinutes: string | null;
    CancelCode: string | null;
    EstDepControlUTCDateTime: string | null;
    FocusFlightIndicator: string | null;
    ArrivalVariance: string | null;
    BlockVariance: string | null;
    DepartureTarmacDelay: string | null;
    ArrivalTarmacDelay: string | null;
    DiversionTarmacDelay: string | null;
    ActualOUTFuelOnBoard: string | null;
    ActualOFFFuelOnBoard: string | null;
    ActualOnFuelOnBoard: string | null;
    ActualInFuelOnBoard: string | null;
    ActualFuelBurn: string | null;
    PlannedOUTFuelOnBoard: string | null;
    PlannedOFFFuelOnBoard: string | null;
    PlannedOnFuelOnBoard: string | null;
    PlannedInFuelOnBoard: string | null;
    PlannedFuelBurn: string | null;
    MealService: string | null;
    Hub: string | null;
    LandingRestriction: string | null;
    HeadStartFlight: string | null;
    SpecialFlight: string | null;
    ScheduledTaxiOutInMinutes: string | null;
    ScheduledTaxiInInMinutes: string | null;
    STISetByUserInMinutes: string | null;
    STOSetByUserInMinutes: string | null;
    CrewTracFlightNumber: string | null;
    DiversionReasonCode: string | null;
    OnwardFlightNumber: string | null;
    OnwardFlightDate: string | null;
    RouteIndex: string | null;
    FlightLegReason: string | null;
}


interface FlightLegReason {
    DelayReasonMinutes: string | null;
    DelayUpdateDateTime: string;
    ReasonType: string;
    ReasonSubType: string | null;
    ReasonCode: string;
    ReasonCodeDesc: string;
    ReasonCategory: string;
}

interface ActualFlightLeg {
    DepStnCode: string;
    PrevTailNumber: string | null;
    SourceSystemName: string;
    DispatchDeskId: string;
    PrevAcftRegNumber: string | null;
    PrevEqpmtType: string | null;
    ActFlightType: string;
    LocalArrDateTime: string;
    SourceFlightLegSeqNumber: string | null;
    LocalDepDateTime: string;
    ArrStnCode: string;
    LocalDepDate: string;
    ServiceType: string;
    ActFlightLegSeqNumber: string;
    EstLocalPubDepDateTime: string;
    DepGate: string | null;
    TailNumber: string;
    SourceMessageType: string;
    SourceFlightLegId: string;
    EventCorrelationId: string;
    Atcstatus: string;
    ArrProgress: string;
    DepProgress: string;
    BaggageCarousel: string | null;
    SchFlightType: string;
    SourceUpdateUserId: string;
    ArrTerminal: string;
    SourceMessageDateTime: string;
    AcftRegNumber: string;
    LegStatus: string;
    SourceUpdateDateTime: string;
    EstLocalPubArrDateTime: string;
    ActLocalOutDateTime: string | null;
    EqpmtType: string;
    OpnlSuffix: string | null;
    EstLocalOpnlDepDateTime: string | null;
    EstLocalOpnlArrDateTime: string | null;
    EstDepControlUtcdateTime: string | null;
    ControlledArrUtcdateTime: string | null;
    PreviousDepGate: string | null;
    ArrGate: string | null;
    EstLocalOffDateTime: string | null;
    ActLocalOffDateTime: string | null;
    ActLocalOnDateTime: string | null;
    ActLocalInDateTime: string | null;
    CompartmentId: string;
    IropstatusCode: string | null;
    IropflightType: string | null;
    ReplacedFromSourceFlightLegSeqNumber: string | null;
    Fitindicator: string | null;
    ActualOUTFuelOnBoard: string | null;
    ActualOFFFuelOnBoard: string | null;
    ActualOnFuelOnBoard: string | null;
    ActualInFuelOnBoard: string | null;
    ActualFuelBurn: string | null;
    PlannedOUTFuelOnBoard: string | null;
    PlannedOFFFuelOnBoard: string | null;
    PlannedOnFuelOnBoard: string | null;
    PlannedInFuelOnBoard: string | null;
    PlannedFuelBurn: string | null;
    AircraftRouting: string | null;
    MealService: string | null;
    Hub: string | null;
    LandingRestriction: string | null;
    HeadStartFlight: string | null;
    SpecialFlight: string | null;
    ScheduledTaxiOutInMinutes: string | null;
    ScheduledTaxiInInMinutes: string | null;
    STISetByUserInMinutes: string | null;
    STOSetByUserInMinutes: string | null;
    CrewTracFlightNumber: string | null;
    DepTerminal: string;
    ETOPSIndicator: string | null;
    OpnlSufx: string | null;
    SourceSystemStatus: string;
    EstLocalOnDatetime: string | null;
    FlightLegReason: FlightLegReason[];
}

interface ScheduledFlightLeg {
    SchOnwardFlightDate: string;
    SecureFlightIndicator: string;
    SourceSystemName: string;
    DomesticArrStnIndictor: string;
    SchAcftConfig: string;
    OnTimePerfValue: string;
    DomesticDepStnIndictor: string;
    ServiceType: string;
    SchAcftOwner: string;
    SourceMessageType: string;
    FlightType: string;
    ScheduledGroundMinutes: string;
    SchOnwardFlightCarrierCode: string;
    Status: string;
    SchOnwardFlightNumber: string;
    SchLocalDepDate: string;
    SchLocalArrDateTime:string;
    RouteNumber:string;
    RouteIndex:string;
    SchOnwardFlightAcftRotationLayover:string;
    ScheduledBlockMinutes:string;
    SchDepStnCode:string;
    ArrTerminal: string;
    DepTerminal: string;
    SchTrafficRestrictionCode:string | null;
    OriginatorFlightIndicator:string | null;
    RON: string | null;
    SourceMessageDateTime: string;
    SchArrStnCode: string;
    FleetLaunchFlightIndicator: string;
    OnTimePerfPeriod: string;
    SourceUpdateDateTime: string;
    EqpmtType: string;
    FlightLegSeqNumber: string;
    AdditionalComment: string | null;
    MarketingFlightSegment: MarketingFlightSegment[];
    CurrentFlightLeg: CurrentFlightLeg;
    ActualFlightLeg: ActualFlightLeg[];
}

export interface FlightDataInterface {
    id: string;
    partition_key: string;
    DocumentType: string;
    FlightCallSign:string;
    FlightCarrierCode: string;
    FlightNumber: string;
    FlightOriginStnCode: string;
    FlightDestStnCode: string;
    FlightOriginLocalDate: string;
    EventSubType: string;
    EntitySeqNumber: string;
    ChangeIndicator: ChangeIndicator[];
    ScheduledFlightLeg: ScheduledFlightLeg[];
}